Move update_analyses to new module
